requirements.txt
deepctr==0.9.0
pandas==1.3.4
scikit_learn==1.0.2
tensorflow_cpu==2.5.0
